requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
distlib==0.3.4
Django==4.0.4
filelock==3.6.0
Flask==2.1.1
frozenlist==1.3.0
future==0.18.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
line-bot-sdk==2.2.1
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.22.3
pipenv==2022.4.21
platformdirs==2.5.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
Werkzeug==2.1.1
yarl==1.7.2
